refactored coping cards
